+++ src/darker/tests/helpers.py
@@ -1,0 +1,8 @@
+"""Helper functions for unit tests"""
+
+from typing import Any, Dict
+
+
+def filter_dict(dct: Dict[str, Any], filter_key: str) -> Dict[str, Any]:
+    """Return only given keys with their values from a dictionary"""
+    return {key: value for key, value in dct.items() if key == filter_key}